move converter check function
