+++ specparam/params/__init__.py
@@ -1,1 +1,4 @@
 """Sub-module for functionality related to parameters and parameter conversions."""
+
+from .converter import AperiodicParamConverter, PeriodicParamConverter
+from .check import check_converters
