+++ specparam/params/definitions.py
@@ -176,19 +176,3 @@ def get_converter(component, parameter, converter):
         converter = NULL_CONVERTERS[component]
 
     return converter
-
-
-def check_converters(component):
-    """Check the set of parameter converters that are available.
-
-    Parameters
-    ----------
-    component : {'aperiodic', 'periodic'}
-        Which component to check available parameter converters for.
-    """
-
-    print('Available {:s} converters:'.format(component))
-    for param, convs in CONVERTERS[component].items():
-        print(param)
-        for label, converter in convs.items():
-            print('    {:10s}    {:s}'.format(converter.name, converter.description))
